refactor: collapse format barrel into format/index.ts
